requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
celery[redis]==5.3.6
redis==5.0.1
qdrant-client==1.7.3
sentence-transformers==2.3.1
python-doctr[torch]==0.8.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
openai==1.12.0
pymupdf==1.23.21
structlog==24.1.0
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
slowapi==0.1.9
aiofiles==23.2.1
